chore: extract shared logic between personal and space editor
